added a template and ASL feature
added a post template for reference as well as a feature to include an optional video of the post signed in ASL.

If the notice on the post to view the ASL video is closed, there will be a cookie planted that expires in 30 days. In other words, it will be 30 days until the site visitor sees the message/banner again.